Separate different screens into activities. Add preferences.
Rather than re-use the main activity screen for two different purposes, I've separated
the main activity and the activity that responds to a broadcast from c:geo.  In doing
this, I've created a new layout as well as some icons for those layouts.

Added a menu bar item for settings. Will eventually put the open source notices there.
The only item on preferences at the moment is to automatically send the cache to the
watch, but this flow isn't perfect. Some cleanup to do there to use the lifecycle the
right way.